+++ src/fileSystemProvider.ts
@@ -11,7 +11,8 @@ import {
 	FileMaps,
 	buildMappedFiles
 } from './repl';
-import {deviceTreeProvider, isPathExist, monocleFolder,screenFolder} from './extension';
+import { startFirmwareUpdate } from './update';
+import {deviceTreeProvider, isPathExist, monocleFolder,configScreenReadUpdate} from './extension';
 import { isConnected } from './bluetooth';
 
 export class File extends vscode.TreeItem implements vscode.FileStat {
@@ -308,25 +309,30 @@ export class ScreenProvider implements vscode.TreeDataProvider<vscode.TreeItem>{
 	private async getAllScreens(){
 		if(vscode.workspace.workspaceFolders){
 			const rootUri = vscode.workspace.workspaceFolders[0].uri;
-			const localFiles = vscode.Uri.joinPath(rootUri,monocleFolder);
+			// const localFiles = vscode.Uri.joinPath(rootUri,screenFolder);
 			
-			if(await isPathExist(localFiles)){
-				const screenFiles = new vscode.RelativePattern(localFiles, screenFolder+'/*_screen.py');
-				const filesFound = await vscode.workspace.findFiles(screenFiles);
-				if(filesFound.length===0){
-					return [];
-				}else{
-					return filesFound.map(f=>{
-						let filename = path.posix.basename(f.path);
-						let newitem = new vscode.TreeItem(filename.replace('_screen.py',''),vscode.TreeItemCollapsibleState.None);
+			if(await isPathExist(rootUri)){
+				// const screenFiles = new vscode.RelativePattern(rootUri, '*_screen.py');
+				let finalList:vscode.TreeItem[]=[];
+				let screensObjs = await configScreenReadUpdate();
+				for (const [key, value] of Object.entries(screensObjs)) {
+					try {
+						let uri = vscode.Uri.joinPath(rootUri,value.filePath);
+							if(await isPathExist(uri)){
+						let newitem = new vscode.TreeItem(key.replace('.py',''),vscode.TreeItemCollapsibleState.None);
 						newitem.iconPath = vscode.ThemeIcon.File;
-						newitem.command =  {command:"brilliant-ar-studio.editUIEditor",title:"Edit In UI Editor",arguments:[{"uri":f,"name":filename.replace('_screen.py','')}]};
-						return newitem;
-					});
+						newitem.command =  {command:"brilliant-ar-studio.editUIEditor",title:"Edit In UI Editor",arguments:[{"uri":uri,"name":key.replace('.py','')}]};
+						finalList.push(newitem);
+							}
+					} catch (error) {
+						console.log(error);
+					}
+				
 				}
+				return finalList;
 			}else{
 				
-				vscode.window.showWarningMessage("Project not initialized!");
+				// vscode.window.showWarningMessage("Project not initialized!");
 				return [];
 			}
 		}else{
@@ -335,3 +341,164 @@ export class ScreenProvider implements vscode.TreeDataProvider<vscode.TreeItem>{
 		
 	}
 }
+
+
+export class DeviceInfoProvider implements vscode.WebviewViewProvider{
+
+	private _view?: vscode.WebviewView;
+	private _currentInfo:any;
+	constructor(
+		private readonly _extensionUri: vscode.Uri,
+	) {
+		this._extensionUri = _extensionUri;
+	 }
+	public updateValues(data:object){
+		this._view?.webview.postMessage(data);
+		this._currentInfo= data;
+	}
+	
+
+	private _setWebviewMessageListener(webview: vscode.Webview) {
+		webview.onDidReceiveMessage(
+		  (message: any) => {
+			switch (message) {
+				case "fpgaUpdate":
+					vscode.commands.executeCommand('brilliant-ar-studio.fpgaUpdate');
+					break;
+				case "customFpga":
+					vscode.commands.executeCommand('brilliant-ar-studio.fpgaUpdateCustom');
+					break;
+				case "firmwareUpdate":
+					startFirmwareUpdate();
+					break;
+				default:
+					break;
+			}
+		  }
+		);
+	}
+    public  resolveWebviewView (webviewView: vscode.WebviewView,  contextWebview: vscode.WebviewViewResolveContext, _token: vscode.CancellationToken){
+            webviewView.webview.options={
+				enableScripts:true,localResourceRoots: [
+				this._extensionUri,
+			]
+		};
+		let updateScript:string = "";
+		if(this._currentInfo){
+			updateScript = 'updateUi(JSON.parse('+JSON.stringify(this._currentInfo)+'))';
+		}
+			this._view = webviewView;
+            webviewView.webview.html=`<!DOCTYPE html>
+			<html lang="en">
+			<head>
+				<meta charset="UTF-8">
+				<meta name="viewport" content="width=device-width, initial-scale=1.0">
+				<title>Device Status</title>
+				<style>
+				a {
+					color: var(--vscode-button-foreground); background-color: var(--vscode-button-background);
+					box-sizing: border-box;
+					display: flex;
+					width: 100%;
+					padding: 4px;
+					border-radius: 2px;
+					text-align: center;
+					cursor: pointer;
+					justify-content: center;
+					align-items: center;
+					border: 1px solid var(--vscode-button-border,transparent);
+					line-height: 18px;
+					text-decoration: none;
+					max-width: 300px;
+					
+				}
+				a:hover{
+					color: var(--vscode-button-foreground); 
+					background: var(--vscode-button-hoverBackground);
+				}
+				button {
+					width: 80vw;
+					max-width:300px;
+					height: 28px;
+					border-radius:2px;
+					margin-top:0.3rem;
+					margin-bottom:0.6rem;
+					background: var(--vscode-button-background);
+					border:  none;
+					color: var(--vscode-button-foreground);
+					cursor: pointer;
+				}
+				button:hover{
+					background: var(--vscode-button-hoverBackground);
+				}
+				</style>
+			</head>
+			
+			<body>
+				<div>
+				Device: <b id="name"></b><br>
+				MAC address: <b id="macAddress"></b><br>
+				Firmware version: <b id="firmwareVersion"></b><br>
+				<a href="javascript:void(0)" id="firmwareUpdate" style="display:none;margin-top:0.5rem;margin-bottom:0.5rem;"></a>
+				FPGA image: <b id="fpgaVersion"></b><br>
+				<a href="javascript:void(0)" id="fpgaUpdate" style="display:none;margin-top:0.5rem;margin-bottom:0.5rem;"></a>
+				<a href="javascript:void(0)" id="customFpga" style="margin-top:1rem">Custom FPGA</a>
+				</div>
+				<script>
+				// To retain state
+				const vscode = acquireVsCodeApi();
+				// Check if we have an old state to restore from
+				const previousState = vscode.getState();
+				if(previousState){
+					updateUi(previousState)
+				}
+				window.addEventListener('message', async (event) => {
+					updateUi(event.data)
+					
+				})
+				document.getElementById('firmwareUpdate').addEventListener('click',function(){
+					vscode.postMessage("firmwareUpdate")
+				})
+				document.getElementById('fpgaUpdate').addEventListener('click',function(){
+					vscode.postMessage("fpgaUpdate")
+				})
+				document.getElementById('customFpga').addEventListener('click',function(){
+					vscode.postMessage("customFpga")
+				})
+				function updateUi(data){
+					vscode.setState(data);
+					if(!data){
+						return;
+					}
+					document.getElementById('name').innerHTML = data.name;
+					document.getElementById('macAddress').innerHTML = data.macAddress.toUpperCase();
+					
+					
+					if(data.firmwareUpdate){
+						document.getElementById('firmwareUpdate').innerHTML = 'Update to'+data.firmwareUpdate;
+						document.getElementById('firmwareUpdate').style.display = 'block';
+						document.getElementById('firmwareVersion').innerHTML = data.firmwareVersion;
+					}else{
+						document.getElementById('firmwareVersion').innerHTML = data.firmwareVersion + '(latest)';
+						document.getElementById('firmwareUpdate').style.display = 'none';
+						document.getElementById('firmwareUpdate').innerHTML = "(Latest)";
+					}
+					if(data.fpgaUpdate){
+						document.getElementById('fpgaUpdate').innerHTML = 'Update to '+data.fpgaUpdate;
+						document.getElementById('fpgaVersion').innerHTML = data.fpgaVersion;
+						document.getElementById('fpgaUpdate').style.display = 'block';
+					}else{
+						document.getElementById('fpgaUpdate').style.display = 'none';
+						document.getElementById('fpgaVersion').innerHTML = data.fpgaVersion+ '(latest)';
+						document.getElementById('fpgaUpdate').innerHTML = "(Latest)";
+					}
+				}
+				${updateScript}
+				</script>
+			</body>
+			</html>`;
+			this._setWebviewMessageListener(webviewView.webview);
+			setTimeout(()=>this.updateValues(this._currentInfo),500);
+    }
+
+};