+++ src/nordicdfu.ts
@@ -20,7 +20,7 @@ export async function startNordicDFU() {
 
         await transferFile(files.dat, 'init');
         await transferFile(files.bin, 'image');
-
+        await new Promise(r => setTimeout(r, 500));
         outputChannel.appendLine('Leaving nRF52 DFU');
         return Promise.resolve("completed");
     } catch (error) {
@@ -200,7 +200,7 @@ async function transferFile(file:any, type:any) {
 
         if (returnedCrc !== chunkCrc) {
              vscode.window.showErrorMessage('CRC mismatch after sending this chunk. Expected: ' + chunkCrc);
-             return Promise.resolve('')
+             return Promise.reject('');
         }
 
         // Execute command